pyproject.toml
[tool.poetry]
name = "homework29drf"
version = "0.1.0"
description = ""
authors = ["Vadim <niras.sp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
pillow = "^9.3.0"
djangorestframework = "^3.14.0"
pyjwt = "^2.6.0"
djangorestframework-simplejwt = "^5.2.2"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
